+++ packages/cropper/src/hooks.tsx
@@ -1,17 +1,12 @@
 import { useRef, useState, RefObject } from "react";
 
 import { exportImage } from "./utils";
-import { Crop } from "./types";
 
-export function useCropper(
-  src: string,
-  crop: Crop,
-  otherProps?: {
-    quality?: number;
-    onSuccess?: (image: string) => void;
-    onError?: (error: Error) => void;
-  }
-): [
+export function useCropper(otherProps?: {
+  quality?: number;
+  onSuccess?: (image: string) => void;
+  onError?: (error: Error) => void;
+}): [
   RefObject<HTMLDivElement | null>,
   {
     result: string | null;
@@ -27,7 +22,30 @@ export function useCropper(
 
   const cropIt = async () => {
     try {
+      let x = 0;
+      let y = 0;
+      let scale = 1;
+
+      let src = null;
+
       if (!ref.current) return;
+
+      if (!src) {
+        src = ref.current.dataset.cropperSrc;
+      }
+
+      /**
+       * Updating the crop from the data attributes
+       */
+      x = Number(ref.current.dataset.cropperX || 0);
+      y = Number(ref.current.dataset.cropperY || 0);
+      scale = Number(ref.current.dataset.cropperScale || 1);
+
+      if (!src) {
+        throw new Error("No src provided");
+      }
+
+      console.log(x, y, scale);
 
       setIsCropping(true);
 
@@ -35,9 +53,9 @@ export function useCropper(
         imageSrc: src,
         cropWidth: ref.current.getBoundingClientRect().width,
         cropHeight: ref.current.getBoundingClientRect().height,
-        x: crop.x,
-        y: crop.y,
-        scale: crop.scale,
+        x,
+        y,
+        scale,
         scaleTheImage: otherProps?.quality || 1,
       });