+++ apps/cropper/CHANGELOG.md
@@ -1,5 +1,19 @@
 # web
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @zuude-ui/cropper@0.0.5
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @zuude-ui/cropper@0.0.4
+
 ## 0.0.3
 
 ### Patch Changes