+++ apps/test/CHANGELOG.md
@@ -1,5 +1,19 @@
 # test
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @zuude-ui/cropper@0.0.5
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @zuude-ui/cropper@0.0.4
+
 ## 0.1.1
 
 ### Patch Changes