+++ packages/cropper/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @zuude-ui/cropper
 
+## 0.0.5
+
+### Patch Changes
+
+- No need to pass src to useCropper hook
+  Disabled the gesture if only crop is provided in the props without onCropChange
+  Added isolate CSS property to cropper container
+
+## 0.0.4
+
+### Patch Changes
+
+- Added disable prop to cropper
+  No more crop scale needed for cropping the image, because I am storing the crop in the cropper container as data attributes
+  Updating the crop when crop prop changes
+
 ## 0.0.3
 
 ### Patch Changes